test: focus Prime KYT regression coverage
